🚨 break change: refactor supabase folder

+++ src/app/(private)/(app)/challenge/[challengeSlug]/@tabContent/components/CommentsList/useCommentsList.ts
@@ -2,16 +2,17 @@
 
 import { useEffect, useState } from 'react'
 import { useRouter } from 'next/navigation'
-import useSWR from 'swr'
 
-import { Comment } from '@/@types/comment'
-import { Order } from '@/@types/order'
+import { Comment } from '@/@types/Comment'
 import { PopoverMenuButton } from '@/app/components/PopoverMenu'
 import { useAuthContext } from '@/contexts/AuthContext/hooks/useAuthContext'
 import { useToast } from '@/contexts/ToastContext'
 import { useApi } from '@/services/api'
+import type { Order } from '@/services/api/types/Order'
+import { useCache } from '@/services/cache'
 import { useChallengeStore } from '@/stores/challengeStore'
-import { ERRORS, ROUTES } from '@/utils/constants'
+import { APP_ERRORS, ROUTES } from '@/utils/constants'
+import { CACHE } from '@/utils/constants/cache'
 
 type Sorter = 'date' | 'upvotes'
 
@@ -59,27 +60,30 @@ export function useCommentsList(canShowComments: boolean) {
     data: comments,
     error,
     isLoading,
-    mutate: refetchComments,
-  } = useSWR(
-    () =>
-      `/comments?challenge_id=${challenge?.id}&user_id=${user?.id}&sorter=${sorter}$order=${order}`,
-    getFilteredComments
-  )
+    refetch,
+  } = useCache({
+    key: CACHE.keys.comments,
+    fetcher: getFilteredComments,
+    dependencies: [challenge?.id, user?.id, sorter, order],
+  })
 
   if (error) {
     console.log(error)
-    throw new Error(ERRORS.comments.failedlistFetching)
+    throw new Error(APP_ERRORS.comments.failedlistFetching)
   }
 
   async function handleDeleteComment(commentId: string) {
-    if (!user) return
+    if (!user || !comments) return
 
     try {
       await api.deleteComment(commentId, user.id)
-      refetchComments()
+      refetch()
     } catch (error) {
       console.error(error)
-      toast.show(ERRORS.comments.failedDeletion, { type: 'error', seconds: 5 })
+      toast.show(APP_ERRORS.comments.failedDeletion, {
+        type: 'error',
+        seconds: 5,
+      })
     }
   }
 
@@ -88,22 +92,21 @@ export function useCommentsList(canShowComments: boolean) {
   }
 
   async function handlePostComment(userComment: string) {
-    if (!challenge || !user) return
+    if (!challenge || !user || !comments) return
 
     try {
       await api.postComment(
         {
-          challenge_id: challenge.id,
+          challengeId: challenge.id,
           content: userComment,
-          parent_comment_id: null,
-          created_at: new Date(),
+          parentCommentId: null,
         },
         user.id
       )
-      await refetchComments()
+      refetch()
     } catch (error) {
       console.error(error)
-      toast.show(ERRORS.comments.failedPost, { type: 'error' })
+      toast.show(APP_ERRORS.comments.failedPost, { type: 'error' })
     } finally {
       setUserComment('')
     }
